requirements.txt
Flask==3.1.0
mysql-connector-python
pandas
scikit-learn
yfinance
ta
joblib
numpy
